another batch of fixes

+++ code/game/gamemodes/meteor/meteors.dm
@@ -124,7 +124,8 @@ GLOBAL_LIST_INIT(meteorsC, list(/obj/effect/meteor/dust)) //for space dust event
 	if(istype(loc, /obj/effect/falling_meteor))
 		var/obj/effect/falling_meteor/holder = loc
 		holder.contained_meteor = null
-		qdel(holder)
+		if(!QDELETED(holder))
+			qdel(holder)
 	. = ..()
 
 /obj/effect/meteor/Initialize(mapload, target)
